fix: resolve coordinator package compilation errors
- Update configuration types to match actual package APIs:
  * parser.Config → parser.ParserConfig
  * planner.Config → planner.PlannerConfig
  * executor.Config → executor.ExecutorConfig
- Create funcEventHandler wrapper to implement EventHandler interface properly
- Fix atomic operation type mismatches:
  * ResourceUsage.GoroutineCount: int → int64
  * WorkerPool.Active: int → int32
- Add missing domain package import to coordinator.go
- Update method signatures for component factories:
  * Use planner.NewExecutionPlanner instead of planner.NewPlanner
  * Add eventBus parameter to executor.NewExecutor and emitter.NewEmitter calls
  * Replace non-existent DefaultConfig functions with inline configuration
- Fix ParserAdapter to use config instead of non-functional parser instance
- Update event handler registration to use proper EventHandler interface
- Fix handler invocation to use Handle method instead of direct function call

The coordinator package now compiles successfully and integrates properly
with the updated parser, planner, executor, and emitter package APIs.

+++ pkg/coordinator/resource_pool.go
@@ -292,7 +292,7 @@ func (r *ConcreteResourcePool) updateResourceUsage() {
 		totalGoroutines += pool.Active
 	}
 	
-	r.resourceUsage.GoroutineCount = int(totalGoroutines)
+	r.resourceUsage.GoroutineCount = int64(totalGoroutines)
 	
 	// Update memory usage (simulated)
 	r.resourceUsage.CurrentMemoryUsage = int64(len(r.workerPools)) * 1024 * 1024 // 1MB per pool
